+++ src/numpy/operators.h
@@ -27,7 +27,7 @@ class Array;
 
 void initialize(int32_t argc, char** argv);
 
-std::shared_ptr<Array> array(std::vector<int64_t> shape, LegateTypeCode type);
+std::shared_ptr<Array> array(std::vector<size_t> shape, LegateTypeCode type);
 
 std::shared_ptr<Array> abs(std::shared_ptr<Array> input);
 
@@ -35,7 +35,7 @@ std::shared_ptr<Array> add(std::shared_ptr<Array> rhs1, std::shared_ptr<Array> r
 
 std::shared_ptr<Array> negative(std::shared_ptr<Array> input);
 
-std::shared_ptr<Array> random(std::vector<int64_t> shape);
+std::shared_ptr<Array> random(std::vector<size_t> shape);
 
 }  // namespace numpy
 }  // namespace legate
